# Remove `_ : Type` from packed classes