… mathcomp/algebra/countalg.v
… mathcomp/algebra/finalg.v
… mathcomp/algebra/ssralg.v
… mathcomp/algebra/ssrnum.v
… mathcomp/fingroup/fingroup.v
… mathcomp/ssreflect/choice.v
… mathcomp/ssreflect/fintype.v
… mathcomp/ssreflect/generic_quotient.v