This PR's purpose is to discuss if it is time to make this change and get some feedback. I did things rather mechanically here, so there might be some room for further improvements.

This increases performance 10% - 15% for Coq v8.6.1 - v8.9.dev.
Tested on a Debain-based 16-core build server and a Macbook Pro laptop with 2,3 GHz Intel Core i5.
It seems Coq at some point fixed the problem `_ : Type` was supposed to solve.

#### Table 1. Debian build server (16 virtual cores), make -j8, real time

|                                        | Compilation time, old (967088a6f87) | Compilation time, new | Speedup |
| -------------------------------------- | ----------------------------------- | --------------------- | ------- |
| Coq 8.6.1                              | 10min 33s                           | 9min 10s              | 15%     |
| Coq 8.7.2                              | 10min 12s                           | 8min 50s              | 15%     |
| Coq 8.8.2                              | 9min 39s                            | 8min 32s              | 13%     |
| Coq 8.9.dev (coq commit 05d827c800544) | 9min 12s                            | 8min 16s              | 11%     |
| Coq 8.10+alpha (coq commit 82eacb323837ac | 6min 33s | 6min 24s | 2.3% |

#### Table 2. Debian build server (16 virtual cores), non-parallel make, real time, Coq 8.8.2

|           | old   | new   | speedup |
| --------- | ----- | ----- | ------- |
| ssreflect | 39s   | 39s   |      0% |
| fingroup  | 54s   | 52s   |    3.8% |
| algebra   | 175s  | 161s  |    8.7% |
| solvable  | 186s  | 181s  |    2.8% |
| field     | 445s  | 357s  |   24.6% |
| character | 245s  | 230s  |    6.5% |
| Total:    | 1144s | 1020s |     12% |

#### Table 3. Debian build server (16 virtual cores), non-parallel make, real time, Coq 8.10+alpha (82eacb323837acc73c2a5bf6b22a7169bbb61e6d)

|           | old   | new   | speedup |
| --------- | ----- | ----- | ------- |
| ssreflect | 37s  | 38s  |   -2.6% |
| fingroup  | 49s  | 50s  |     -2% |
| algebra   | 143s | 140s |    2.1% |
| solvable  | 173s | 170s |    1.8% |
| field     | 209s | 190s |     10% |
| character | 220s | 205s |    7.3% |
| Total:    | 831s | 793s |    4.8% |