@thery 2018-11-22
@anton-trunov is this speed-up for the compilation of the library or for Feit-Thompson?