@ejgallego 2018-11-22
I am not sure if this is related, just in case: https://github.com/coq/coq/pull/396