@gares 2018-11-22
CC @maximedenes @ppedrot 

IIRC these copies of the first argument (replicated at the end of the record) were there to make the term comparison (of the kernel) fail early (since it goes right-to-left).

Maybe it is never the case that the records touched by this PR needs to be compared as different by the kernel. Is there a way we can check this assertion? Is there a way to keep the parameter still there but ignore it at term comparison time (to be sure this is the place were we gain speed)?