@anton-trunov 2018-11-22
> @anton-trunov is this speed-up for the compilation of the library or for Feit-Thompson?

@thery This is speedup for the compilation of all the packages in *this* repository, I didn't test `odd-order` repository.
I also tested the speedups for each individual package, please see the updated initial message.