@anton-trunov 2018-11-22
>  I'd be curious to see how this PR fares atop of a recent master containing the aforementioned tweak.

@ppedrot I updated the initial message (see updated table 1 and a new table 3).

Amazing job the Coq devs did! Congrats!