@gares 2018-11-23
So this pr is a 10% speedup on top of the 100% speedup given by the coq patch recently merged, on the field component.

I believe that it would be relevant to apply this pr, even if it was gaining nothing on the performance side, since it cleans up the structures. 10% is the cherry on top.