@CohenCyril 2018-11-23
So before merging, I would like to
1. have a benchmark on odd-order too, and
2. have a regression test on failure cases for kernel comparison that used to be too slow... @ggonthier, @gares did you keep such examples somewhere?