@anton-trunov 2018-11-23
> have a benchmark on odd-order too

It turns out `odd-order` needs a 4-character fix, see [here](https://github.com/anton-trunov/odd-order/commit/10f523ff7d8ed1f52184dd61a56f35410ed18977)) before it can be compiled with this PR.

With that fix, I did a benchmark on the aforementioned laptop with Coq 8.10+alpha (coq/coq@8fb01564fba587142c2471708ff18219f1c64903).
- Old: math-comp/odd-order@10f523ff7d8ed1f52184dd61a56f35410ed18977, mathcomp (967088a6f87405a93ce21971392c58996df8c99f)
`make  906,10s user 13,57s system 99% cpu 15:23,02 total`

- New: math-comp/odd-order@10f523ff7d8ed1f52184dd61a56f35410ed18977, mathcomp (27e3e35be9a6805e126df5b44a9bb5db9a561f5a)
`make  889,57s user 13,55s system 99% cpu 15:06,49 total`

Resulting in 1.86% speedup (probably just noise).

I hope this helps.