@anton-trunov 2018-12-04
Rebased to fix a conflict with 03ad994dfee48e1a7b2b7091c45dfdcf4402f826