@CohenCyril 2018-12-06
Please also add a ChangeLog entry