@anton-trunov 2018-12-06
> Please also add a ChangeLog entry

@CohenCyril Updated. Please let me know if you'd like to expand the entry.